+++ src/main/clojure/clara/rules/durability.clj
@@ -466,7 +466,7 @@
 
 (defn opts->get-alphas-fn [rulebase opts]
   (let [fact-type-fn (:fact-type-fn opts type)
-        ancestors-fn (:ancestors-fn opts ancestors)]
+        ancestors-fn (com/create-ancestors-fn opts)]
     (create-get-alphas-fn fact-type-fn
                           ancestors-fn
                           (:alpha-roots rulebase))))
@@ -555,7 +555,9 @@
    Options for the rulebase semantics that are documented at clara.rules/mk-session include:
 
    * :fact-type-fn
+   * :hash-expr-fn
    * :ancestors-fn
+   * :hierarchy
    * :activation-group-sort-fn
    * :activation-group-fn