+++ src/main/clojure/clara/rules/durability/fressian.clj
@@ -602,17 +602,13 @@
 
     (with-open [^FressianReader rdr (fres/create-reader in-stream :handlers read-handler-lookup)]
       (let [{:keys [rulebase-only?
-                    base-rulebase
-                    forms-per-eval]} opts
+                    base-rulebase]} opts
 
             record-holder (ArrayList.)
             ;; The rulebase should either be given from the base-session or found in
             ;; the restored session-state.
             maybe-base-rulebase (when (and (not rulebase-only?) base-rulebase)
                                   base-rulebase)
-
-            forms-per-eval (or forms-per-eval com/forms-per-eval-default)
-            compiler-cache (get opts :compiler-cache com/default-compiler-cache)
 
             reconstruct-expressions (fn [expr-lookup]
                                       ;; Rebuilding the expr-lookup map from the serialized map:
@@ -629,7 +625,7 @@
                                                           d/clj-struct-holder record-holder]
                                                          (pform/thread-local-binding [d/node-fn-cache (-> (fres/read-object rdr)
                                                                                                           reconstruct-expressions
-                                                                                                          (com/compile-exprs compiler-cache forms-per-eval))]
+                                                                                                          (com/compile-exprs opts))]
                                                                                      (assoc (fres/read-object rdr)
                                                                                             :node-expr-fn-lookup
                                                                                             (.get d/node-fn-cache))))]