+++ CHANGELOG.md
@@ -3,11 +3,11 @@ This is a history of changes to k13labs/clara-rules.
 # 1.4.0-SNAPSHOT
 * update linter config for new macros.
 * `defrule` now defines rules as functions with two arities, no args returns the rule map, and 2 args is the compiled RHS.
+* `defrule` separator `=>` symbol is now optional when there is no LHS expressions
 * `clojure.lang.Fn` now implements `clara.rules.compiler/IRuleSource`, and returns a single rule by invoking the like `(a-rule)`.
 * add built-in support to serialize `clojure.lang.Var` so that a rule handler var can be serialized correctly.
 * add function `clara.rules.compiler/load-rules-from-source` to simplify loading rules
 * add `defhierarchy` macro to define hierarchies of facts allowing to easily establish parent/child relationships.
-* add `defdata` macro to define facts and collections of facts in namespaces allowing to easily embed and insert them during mk-session.
 * rename `clear-ns-productions!` to `clear-ns-vars!` since now there are ns-installed vars that are not productions.
 
 # 1.3.4